fastpi-poject/requirements.txt
# Core FastAPI stack
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic==2.7.1
python-dotenv==1.0.1
dnspython==2.6.1   

# Optional: Testing & Dev Tools
pytest==8.2.2
httpx==0.27.0
ruff==0.4.4

# Optional: Type checking and formatting
mypy==1.10.0
black==24.4.2